feat(EPIC-001): Phase 0 — decimals unification and u128 token balances
Remove CBE constants from lib-types (CBE is a DAO, not a system token).
Unify SOV to 18 decimals everywhere. Widen TokenContract balances,
supply, and all amount parameters from u64 to u128.

Changes:
- lib-types/tokenomics.rs: remove CBE_DECIMALS, CBE_MAX_SUPPLY,
  CBE_TOTAL_SUPPLY_TOKENS (moved to cbe_token.rs as self-contained)
- constants.rs: SOV_TOKEN_DECIMALS 8 → 18, SOV_TOKEN_MAX_SUPPLY u64 → u128
- core.rs: TokenContract balances/supply/amounts all u128
- canonical.rs: bonding curve MAX_SUPPLY defined locally (was CBE_MAX_SUPPLY)
- 37 files updated to propagate u128 through callers

1786 lib-blockchain tests pass, 0 failures.

Refs: #2093, #2094

+++ lib-blockchain/src/contracts/ubi_distribution/core.rs
@@ -361,7 +361,7 @@ impl UbiDistributor {
         } else {
             // Legacy direct path
             let _burned = token
-                .transfer(ctx, citizen, amount)
+                .transfer(ctx, citizen, amount as u128)
                 .map_err(|_| Error::TokenTransferFailed)?;
         }
 
@@ -1114,14 +1114,14 @@ mod tests {
             "Test Token".to_string(),
             "TTK".to_string(),
             8,
-            u64::MAX,
+            u128::MAX,
             false,
             0,
             creator.clone(),
         );
 
         // Mint a large balance to the contract address for transfers
-        let _ = token.mint(contract_address, u64::MAX / 2);
+        let _ = token.mint(contract_address, u128::MAX / 2);
         token
     }
 
@@ -1135,7 +1135,7 @@ mod tests {
             "Test Token".to_string(),
             "TTK".to_string(),
             8,
-            u64::MAX,
+            u128::MAX,
             false,
             0,
             creator.clone(),
